Simplify move operations
They have explicit copy methods because a copy constructor isn't relevant in all cases. In a triple register, it is still four words wide but only operates on the first three. Copy construction would not be straight forward in what it is doing internally.

This is easy to fix in the future though...